requirements.txt
tensorflow
numpy
Pillow
matplotlib
scikit-learn
opencv-python